% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malat1_function.R
\name{define_malat1_threshold}
\alias{define_malat1_threshold}
\title{Define MALAT1 Expression Threshold}
\arguments{
\item{counts}{A numeric vector of normalized MALAT1 counts.}

\item{bw}{A numeric value specifying the bandwidth for density estimation. Default is 0.1.}

\item{lwd}{A numeric value specifying the line width for the red line in the histogram. Default is 2.}

\item{breaks}{An integer specifying the number of bins for the histogram. Default is 100.}

\item{chosen_min}{A numeric value specifying the minimum threshold for the MALAT1 peak. Default is 1.}

\item{smooth}{A numeric value specifying the smoothing parameter for the smooth spline. Default is 1.}

\item{abs_min}{A numeric value specifying the absolute minimum for the local minima. Default is 0.3.}

\item{rough_max}{A numeric value specifying the rough maximum for the local maxima. Default is 2.}
}
\value{
A numeric value representing the calculated MALAT1 threshold.
}
\description{
This function calculates a minimum threshold value for MALAT1 expression.
Given a vector of normalized MALAT1 expression values, this function identifies
an optimal threshold to aid in filtering empty droplets in single-cell RNA sequencing (scRNA-seq) data.
}
\examples{
counts <- rnorm(1000, mean = 1, sd = 0.5)
define_malat1_threshold(counts)
}
