% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_barcodes}
\alias{example_barcodes}
\title{Example cell barcodes}
\format{
A character vector of cell barcodes.
}
\source{
Internal example dataset.
}
\usage{
example_barcodes
}
\description{
A character vector containing cell barcodes corresponding to the values in
\code{\link{example_counts}}.
}
\keyword{datasets}
