% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{Example MALAT1 count data}
\format{
A numeric vector containing MALAT1 counts for individual cells.
}
\source{
Internal example dataset.
}
\usage{
example_counts
}
\description{
A vector of MALAT1 expression counts from a representative single-cell
RNA-sequencing dataset. These counts can be used as input to
\code{\link{malat1_thres_raw}} or \code{\link{malat1_thres_filtered}}.
}
\keyword{datasets}
