% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10X.R
\name{malat1_read_10X_h5_raw}
\alias{malat1_read_10X_h5_raw}
\title{Filter cells based on MALAT1 expression from a raw CellRanger h5 file.}
\usage{
malat1_read_10X_h5_raw(
  raw_cellranger_h5,
  use.names = TRUE,
  unique.features = TRUE,
  malat1_name = "MALAT1",
  min_malat1 = 4,
  print_plots = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{raw_cellranger_h5}{CellRanger output: raw_feature_bc_matrix_h5.h5.}

\item{use.names}{Row names with feature names rather than ID numbers (default TRUE).}

\item{unique.features}{Logical to make feature names unique (default TRUE)}

\item{malat1_name}{Gene symbol used for MALAT1 (default "MALAT1"); use "Malat1" for mouse or Ensembl gene names if required}

\item{min_malat1}{The minimum MALAT1 count required for a value to be included in the mixture model fitting.}

\item{print_plots}{Logical indicating whether or not to print MALAT1 histogram with model fit.}

\item{return_plots}{Logical indicating whether or not to return MALAT1 histogram with model fit.}
}
\description{
Filter cells based on MALAT1 expression from a raw CellRanger h5 file.
}
