% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10X.R
\name{malat1_read_10X_raw}
\alias{malat1_read_10X_raw}
\title{Filter cells based on MALAT1 expression from a raw CellRanger folder.}
\usage{
malat1_read_10X_raw(
  raw_cellranger_folder,
  unique.features = TRUE,
  malat1_name = "MALAT1",
  min_malat1 = 4,
  print_plots = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{raw_cellranger_folder}{Folder containing barcodes.tsv, genes.tsv, and matrix.mtx}

\item{unique.features}{Logical to make feature names unique (default TRUE)}

\item{malat1_name}{Gene symbol used for MALAT1 (default "MALAT1"); use "Malat1" for mouse or Ensembl gene names if required}

\item{min_malat1}{The minimum MALAT1 count required for a value to be included in the mixture model fitting.}

\item{print_plots}{Logical indicating whether or not to print MALAT1 histogram with model fit.}

\item{return_plots}{Logical indicating whether or not to return MALAT1 histogram with model fit.}
}
\description{
Filter cells based on MALAT1 expression from a raw CellRanger folder.
}
