% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malat1_threshold.R
\name{malat1_thres_filtered}
\alias{malat1_thres_filtered}
\title{Filter cells based on MALAT1 expression from a filtered dataset of cells that may include empty droplets}
\usage{
malat1_thres_filtered(
  counts,
  barcodes,
  log_transform = TRUE,
  min_malat1 = 1,
  breaks = 100,
  print_plots = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{counts}{A vector of MALAT1 counts.}

\item{barcodes}{A vector of cell barcodes.}

\item{log_transform}{Whether to apply a log1p transformation to MALAT1 counts before fitting the mixture model.}

\item{min_malat1}{The minimum MALAT1 count required for a value to be included in the mixture model fitting.}

\item{breaks}{Number of bins used to generate the MALAT1 histogram.}

\item{print_plots}{Logical indicating whether or not to print MALAT1 histogram with model fit.}

\item{return_plots}{Logical indicating whether or not to return MALAT1 histogram with model fit.}
}
\description{
Filter cells based on MALAT1 expression from a filtered dataset of cells that may include empty droplets
}
\examples{

data(example_counts)
data(example_barcodes)

result <- malat1_thres_filtered(
  example_counts,
  example_barcodes,
  print_plots = TRUE
)

result$threshold
result$cells[1:5]
}
