% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malat1_threshold.R
\name{malat1_thres_raw}
\alias{malat1_thres_raw}
\title{Filter cells based on MALAT1 expression from a raw count matrix containing cells and droplets}
\usage{
malat1_thres_raw(
  raw_counts,
  barcodes,
  min_malat1 = 4,
  print_plots = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{raw_counts}{A vector of MALAT1 counts.}

\item{barcodes}{A vector of cell barcodes.}

\item{min_malat1}{The minimum MALAT1 count required for a value to be included in the mixture model fitting.}

\item{print_plots}{Logical indicating whether or not to print MALAT1 histogram with model fit.}

\item{return_plots}{Logical indicating whether or not to return MALAT1 histogram with model fit.}
}
\description{
Filter cells based on MALAT1 expression from a raw count matrix containing cells and droplets
}
\examples{

data(example_counts)
data(example_barcodes)

result <- malat1_thres_raw(
  example_counts,
  example_barcodes,
  print_plots = TRUE
)

result$threshold
result$cells[1:5]
}
